move getFacets from Path to Part
